requirements.txt
grpcio-tools==1.48.2
grpcio==1.48.2
protobuf==3.19.6
pypiwin32
pywin32
Cython==0.29.32
setuptools==59.6.0
rich~=12.6.0
